refactor: index.ts for config
