+++ packages/apps/reputation-oracle/server/src/config/index.ts
@@ -1,1 +1,14 @@
 export * from './env-schema';
+
+export { AuthConfigService } from './auth-config.service';
+export { DatabaseConfigService } from './database-config.service';
+export { EmailConfigService } from './email-config.service';
+export { HCaptchaConfigService } from './hcaptcha-config.service';
+export { KycConfigService } from './kyc-config.service';
+export { NDAConfigService } from './nda-config.service';
+export { PGPConfigService } from './pgp-config.service';
+export { ReputationConfigService } from './reputation-config.service';
+export { S3ConfigService } from './s3-config.service';
+export { ServerConfigService } from './server-config.service';
+export { Web3ConfigService } from './web3-config.service';
+export { SlackConfigService } from './slack-config.service';
