+++ packages/apps/reputation-oracle/server/src/integrations/hcaptcha/hcaptcha.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { ethers } from 'ethers';
 import { firstValueFrom } from 'rxjs';
 
-import { HCaptchaConfigService } from '../../config/hcaptcha-config.service';
+import { HCaptchaConfigService } from '../../config';
 import logger from '../../logger';
 
 import {
