+++ packages/apps/reputation-oracle/server/src/modules/abuse/abuse-slack-bot.spec.ts
@@ -3,7 +3,7 @@ import { Test } from '@nestjs/testing';
 import { HttpService } from '@nestjs/axios';
 
 import { createHttpServiceMock } from '../../../test/mock-creators/nest';
-import { SlackConfigService } from '../../config/slack-config.service';
+import { SlackConfigService } from '../../config';
 
 import { AbuseSlackBot } from './abuse-slack-bot';
 import { AbuseDecision } from './constants';
