+++ packages/apps/reputation-oracle/server/src/modules/email/sendgrid.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
 import { MailService } from '@sendgrid/mail';
 
-import { EmailService } from './email.service';
-import { EmailAction, SENDGRID_API_KEY_REGEX } from './constants';
+import { EmailConfigService } from '../../config';
 import logger from '../../logger';
-import { EmailConfigService } from '../../config/email-config.service';
 import Environment from '../../utils/environment';
+import { EmailAction, SENDGRID_API_KEY_REGEX } from './constants';
+import { EmailService } from './email.service';
 
 export const SENDGRID_TEMPLATES = {
   signup: 'd-ca99cc7410aa4e6dab3e6042d5ecb9a3',
