+++ packages/apps/reputation-oracle/server/src/modules/encryption/pgp-encryption.service.spec.ts
@@ -14,8 +14,7 @@ import { faker } from '@faker-js/faker';
 import { Encryption, EncryptionUtils, KVStoreUtils } from '@human-protocol/sdk';
 import { Test } from '@nestjs/testing';
 
-import { PGPConfigService } from '../../config/pgp-config.service';
-import { Web3ConfigService } from '../../config/web3-config.service';
+import { PGPConfigService, Web3ConfigService } from '../../config';
 import {
   generateTestnetChainId,
   mockWeb3ConfigService,
