+++ packages/apps/reputation-oracle/server/src/modules/encryption/pgp-encryption.service.ts
@@ -6,7 +6,7 @@ import {
 } from '@human-protocol/sdk';
 import { Injectable, OnModuleInit } from '@nestjs/common';
 
-import { PGPConfigService } from '../../config/pgp-config.service';
+import { PGPConfigService } from '../../config';
 import logger from '../../logger';
 
 import { Web3Service } from '../web3/web3.service';
