+++ packages/apps/reputation-oracle/server/src/modules/health/health.controller.ts
@@ -9,7 +9,7 @@ import {
 } from '@nestjs/terminus';
 import packageJson from '../../../package.json';
 import { Public } from '../../common/decorators';
-import { ServerConfigService } from '../../config/server-config.service';
+import { ServerConfigService } from '../../config';
 import { PingResponseDto } from './dto/ping-response.dto';
 import Environment from '../../utils/environment';
