+++ packages/apps/reputation-oracle/server/src/modules/kyc/fixtures/index.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { KycConfigService } from '../../../config/kyc-config.service';
+import { KycConfigService } from '../../../config';
 import { KycEntity } from '../kyc.entity';
 import { KycStatus } from '../constants';
