+++ packages/apps/reputation-oracle/server/src/modules/kyc/kyc.service.spec.ts
@@ -4,15 +4,12 @@ import { HttpService } from '@nestjs/axios';
 import { Test } from '@nestjs/testing';
 import { ethers } from 'ethers';
 
-import { KycConfigService } from '../../config/kyc-config.service';
-import { Web3ConfigService } from '../../config/web3-config.service';
-
 import { generateEthWallet } from '../../../test/fixtures/web3';
 import {
   createHttpServiceMock,
   createHttpServiceResponse,
 } from '../../../test/mock-creators/nest';
-
+import { KycConfigService, Web3ConfigService } from '../../config';
 import { KycStatus } from '../kyc/constants';
 import { UserEntity } from '../user';
 import { generateWorkerUser } from '../user/fixtures';
