+++ packages/apps/reputation-oracle/server/src/modules/kyc/kyc.service.ts
@@ -3,8 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { AxiosError } from 'axios';
 import { catchError, firstValueFrom } from 'rxjs';
 
-import { KycConfigService } from '../../config/kyc-config.service';
-import { Web3ConfigService } from '../../config/web3-config.service';
+import { KycConfigService, Web3ConfigService } from '../../config';
 import logger from '../../logger';
 import * as httpUtils from '../../utils/http';
 import { UserEntity } from '../user';
