+++ packages/apps/reputation-oracle/server/src/modules/nda/fixtures/index.ts
@@ -1,6 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-import { NDAConfigService } from '../../../config/nda-config.service';
+import { NDAConfigService } from '../../../config';
 
 export const mockNdaConfigService: Omit<NDAConfigService, 'configService'> = {
   latestNdaUrl: faker.internet.url(),
