+++ packages/apps/reputation-oracle/server/src/modules/qualification/qualification.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { v4 as uuidV4 } from 'uuid';
 
-import { ServerConfigService } from '../../config/server-config.service';
+import { ServerConfigService } from '../../config';
 import logger from '../../logger';
 import { UserRepository, UserStatus } from '../user';
