+++ packages/apps/reputation-oracle/server/src/modules/reputation/reputation.service.spec.ts
@@ -5,8 +5,7 @@ import { faker } from '@faker-js/faker';
 import { EscrowClient } from '@human-protocol/sdk';
 import { Test } from '@nestjs/testing';
 
-import { ReputationConfigService } from '../../config/reputation-config.service';
-import { Web3ConfigService } from '../../config/web3-config.service';
+import { ReputationConfigService, Web3ConfigService } from '../../config';
 
 import {
   generateTestnetChainId,
