+++ packages/apps/reputation-oracle/server/src/modules/storage/storage.service.spec.ts
@@ -6,7 +6,7 @@ import { Test } from '@nestjs/testing';
 import { Client as MinioClient } from 'minio';
 
 import { ContentType } from '../../common/enums';
-import { S3ConfigService } from '../../config/s3-config.service';
+import { S3ConfigService } from '../../config';
 import * as httpUtils from '../../utils/http';
 
 import { PgpEncryptionService } from '../encryption/pgp-encryption.service';
