+++ packages/apps/reputation-oracle/server/src/modules/storage/storage.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import * as Minio from 'minio';
 
 import { ContentType } from '../../common/enums';
-import { S3ConfigService } from '../../config/s3-config.service';
+import { S3ConfigService } from '../../config';
 import logger from '../../logger';
 import * as httpUtils from '../../utils/http';
