+++ packages/apps/reputation-oracle/server/src/modules/webhook/webhook-incoming.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 
 import { BACKOFF_INTERVAL_SECONDS } from '../../common/constants';
 import { isDuplicatedError } from '../../common/errors/database';
-import { ServerConfigService } from '../../config/server-config.service';
+import { ServerConfigService } from '../../config';
 import { EscrowCompletionService } from '../escrow-completion/escrow-completion.service';
 import { calculateExponentialBackoffMs } from '../../utils/backoff';
 import logger from '../../logger';
